requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.23.0
anthropic>=0.39.0
httpx>=0.25.0
python-dotenv>=1.0.0
